Mrge fuel controller tests into test_fuel_controller
