+++ test/test_schedules/test_ignition_controller.cpp
@@ -1,7 +1,6 @@
 #include <unity.h>
 #include "../test_utils.h"
 #include "scheduler.h"
-#include "channel_test_helpers.h"
 
 void test_ignition_controller(void)
 {
