+++ test/test_schedules/test_status_off_to_pending.cpp
@@ -3,7 +3,7 @@
 #include <unity.h>
 #include "../test_utils.h"
 #include "scheduler_fuel_controller.h"
-#include "channel_test_helpers.h"
+#include "../channel_test_helpers.h"
 #include "scheduler_ignition_controller.h"
 
 #define TIMEOUT 1000
